requirements.txt
Flask
hyperliquid-python-sdk
python-dotenv
msgpack
web3